improved level accuracy
there's an interesting bug with accuracy where the iframe element for the user code is slightly different than the iframe level solution when taking puppeteer screenshots. it's weird because these elements are both in the same position and have the exact same size. to counter this bug, the solution images match the exact pixels of the user solution iframe, and now it can finally give up to 100% accuracy